Refactor folder structure of admin views
