+++ src/components/AdminView/AdminNavigation/AdminNavigation.scss
@@ -1,5 +1,5 @@
-@import "../../utils/_colors.scss";
-@import "./margin.scss";
+@import "../../../utils/_colors.scss";
+@import "../margin.scss";
 
 $navWidth: 480px;
