+++ src/components/AdminView/AdminNavigation/AdminNavigationLink.scss
@@ -1,5 +1,5 @@
-@import "../../utils/_colors.scss";
-@import "./margin";
+@import "../../../utils/_colors.scss";
+@import "../margin";
 
 .admin-nav-link {
   list-style-type: none;
