+++ src/components/AdminView/AdminUploadView/AdminUploadConfirmModal.scss
@@ -1,4 +1,4 @@
-@import "../../../../utils/_colors.scss";
+@import "../../../utils/_colors.scss";
 
 .AdminUploadConfirmModal {
   position: absolute;
