+++ src/components/AdminView/AdminUploadView/AdminUploadView.scss
