+++ src/components/AdminView/AdminUploadView/DragAndDrop.scss
@@ -1,4 +1,4 @@
-@import "../../../../utils/_colors.scss";
+@import "../../../utils/_colors.scss";
 
 .DragAndDrop {
   position: relative;
